+++ src/server/server.ts
@@ -5,6 +5,7 @@ import {SVELTE_KIT_DIST_PATH} from '@feltcoop/gro/dist/paths.js';
 import {ApiServer} from './ApiServer.js';
 import {Database} from '../db/Database.js';
 import {toDefaultPostgresOptions} from '../db/postgres.js';
+import {seed} from '../db/seed.js';
 
 export const server = new ApiServer({
 	app: polka(),
@@ -20,7 +21,12 @@ export const server = new ApiServer({
 	},
 });
 
-server.init().catch((err) => {
-	console.error('server.init() failed', err);
-	throw err;
-});
+server
+	.init()
+	.then(async () => {
+		await seed(server);
+	})
+	.catch((err) => {
+		console.error('server.init() failed', err);
+		throw err;
+	});
